resolved merge conflicts from selenide-master
